Create pizza_order_practice
Program to calculate the cost of a pizza order where the price of a Small pizza is $15, Medium pizza is $20, and a Large pizza is $25. If the user wants to add pepperoni, it will add $2 to the bill if the pizza is Small, or $3 if the pizza is Medium or Large. If the user wants extra cheese, it will add $1 to the bill.