Create fizz_buzz.py
The Fizz Buzz coding challenge which prints out the numbers 1 to 100. If a number is cleanly divisible by 3, it'll print the word "Fizz" instead of that number. If a number is cleanly divisible by 5, it'll print the word "Buzz" instead of that number. And if a number is cleanly divisible by 3 and by 5, it'll print the word "FizzBuzz" instead of that number.