Create paint_can_calculator.py
A program which calculates the number of paint cans required to paint a wall, where each can covers 5 square meters. The program asks the user to enter the height and width of the wall and then calculates the area of the wall. It then divides that value by the amount of coverage each can provides (5) and this determines how many cans of paint is required.