Create prime_number_checker.py
A program which checks if a number is a Prime number. A number is Prime only if it is cleanly divisible by 1 and itself. The program takes in a number and uses a for loop to iterate through every number in the range of 2 to the inputted number. It uses a modulus to check if the inputted number is cleanly divisible by the number in the range. If it is, then it is not a Prime number.