Create grading_program.py
A program that loops through a dictionary called student_scores, which contains student names as the key and their grades as the value and creates a new dictionary called student_grades. Student_grades has the student's name as the key and an appropriate message, depending on which grade the student got, as the value.