Create etch_a_sketch.py
Creates a Etch a Sketch program with the Turtle library. The user is able to draw with the turtle, using keys on the keyboard to move it - W for forwards, S for backwards, D to turn right, and A to turn left. The C key can be used to clear the screen and put the turtle back at the starting position.